+++ client/src/api/data.js
@@ -1,0 +1,6 @@
+// Sustainability words
+export const words = ['RenewableEnergy', 'Recycling', 'Biodiversity'];
+
+export const wordsWithCheckBox = words.reduce((allWords, currentWord) => {
+  return { ...allWords, [currentWord]: false };
+}, {});
